Refactor populate.
